All disk option buttons work!

+++ src/main/java/com/example/towersofhanoi/Game.java
@@ -1,5 +1,5 @@
 package com.example.towersofhanoi;
 
 public class Game {
-    public static int disks;
+    public static byte disks;
 }
